+++ scripts/build_solar_thermal_profiles.py
@@ -5,6 +5,7 @@ import atlite
 import pandas as pd
 import xarray as xr
 import numpy as np
+from dask.distributed import Client, LocalCluster
 
 if __name__ == '__main__':
     if 'snakemake' not in globals():
@@ -15,14 +16,9 @@ if __name__ == '__main__':
             clusters=48,
         )
 
-    if 'snakemake' not in globals():
-        from vresutils import Dict
-        import yaml
-        snakemake = Dict()
-        with open('config.yaml') as f:
-            snakemake.config = yaml.safe_load(f)
-        snakemake.input = Dict()
-        snakemake.output = Dict()
+    nprocesses = int(snakemake.threads)
+    cluster = LocalCluster(n_workers=nprocesses, threads_per_worker=1)
+    client = Client(cluster, asynchronous=True)
 
     config = snakemake.config['solar_thermal']
 
@@ -35,18 +31,18 @@ if __name__ == '__main__':
 
     I = cutout.indicatormatrix(clustered_regions)
 
-    for area in ["total", "rural", "urban"]:
+    pop_layout = xr.open_dataarray(snakemake.input.pop_layout)
 
-        pop_layout = xr.open_dataarray(snakemake.input[f'pop_layout_{area}'])
+    stacked_pop = pop_layout.stack(spatial=('y', 'x'))
+    M = I.T.dot(np.diag(I.dot(stacked_pop)))
 
-        stacked_pop = pop_layout.stack(spatial=('y', 'x'))
-        M = I.T.dot(np.diag(I.dot(stacked_pop)))
+    nonzero_sum = M.sum(axis=0, keepdims=True)
+    nonzero_sum[nonzero_sum == 0.] = 1.
+    M_tilde = M / nonzero_sum
 
-        nonzero_sum = M.sum(axis=0, keepdims=True)
-        nonzero_sum[nonzero_sum == 0.] = 1.
-        M_tilde = M / nonzero_sum
+    solar_thermal = cutout.solar_thermal(**config, matrix=M_tilde.T,
+                                            index=clustered_regions.index,
+                                            dask_kwargs=dict(scheduler=client),
+                                            show_progress=False)
 
-        solar_thermal = cutout.solar_thermal(**config, matrix=M_tilde.T,
-                                             index=clustered_regions.index)
-
-        solar_thermal.to_netcdf(snakemake.output[f"solar_thermal_{area}"])
+    solar_thermal.to_netcdf(snakemake.output.solar_thermal)